update logging
增加了日志文件
