+++ loan/log_test.log
@@ -1,0 +1,12 @@
+2019-06-08 17:57:56.711  INFO 9836 --- [main] loan.easyLoan.LoanApplicationTests       : Starting LoanApplicationTests on DESKTOP-EN1OLRT with PID 9836 (started by admin in E:\2019金证实习\easyLoan\loan)
+2019-06-08 17:57:56.712 DEBUG 9836 --- [main] loan.easyLoan.LoanApplicationTests       : Running with Spring Boot v2.1.5.RELEASE, Spring v5.1.7.RELEASE
+2019-06-08 17:57:56.712  INFO 9836 --- [main] loan.easyLoan.LoanApplicationTests       : No active profile set, falling back to default profiles: default
+2019-06-08 17:57:58.640  WARN 9836 --- [main] o.m.s.mapper.ClassPathMapperScanner      : No MyBatis mapper was found in '[loan.easyLoan]' package. Please check your configuration.
+2019-06-08 17:58:00.076  INFO 9836 --- [main] o.s.s.concurrent.ThreadPoolTaskExecutor  : Initializing ExecutorService 'applicationTaskExecutor'
+2019-06-08 17:58:01.133  INFO 9836 --- [main] loan.easyLoan.LoanApplicationTests       : Started LoanApplicationTests in 5.061 seconds (JVM running for 6.501)
+2019-06-08 17:58:01.433 TRACE 9836 --- [main] loan.easyLoan.LoanApplicationTests       : trace logger...
+2019-06-08 17:58:01.434 DEBUG 9836 --- [main] loan.easyLoan.LoanApplicationTests       : debug logger...
+2019-06-08 17:58:01.434  INFO 9836 --- [main] loan.easyLoan.LoanApplicationTests       : info logger...
+2019-06-08 17:58:01.434  WARN 9836 --- [main] loan.easyLoan.LoanApplicationTests       : warn logger...
+2019-06-08 17:58:01.434 ERROR 9836 --- [main] loan.easyLoan.LoanApplicationTests       : error logger...
+2019-06-08 17:58:01.446  INFO 9836 --- [Thread-2] o.s.s.concurrent.ThreadPoolTaskExecutor  : Shutting down ExecutorService 'applicationTaskExecutor'
