+++ loan/src/main/java/loan/easyLoan/LoanApplication.java
@@ -1,13 +1,17 @@
-package loan.easyloan;
+package loan.easyLoan;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
 public class LoanApplication {
 
+
     public static void main(String[] args) {
         SpringApplication.run(LoanApplication.class, args);
+
     }
 
 }
