+++ loan/src/main/java/loan/easyLoan/controller/HelloController.java
@@ -1,4 +1,4 @@
-package loan.easyloan.controller;
+package loan.easyLoan.controller;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
