+++ loan/src/main/java/loan/easyLoan/entity/BorrowMoneyFlow.java
@@ -1,17 +1,14 @@
-package loan.easyloan.entity;
+package loan.easyLoan.entity;
 
 
-public class RepayMoneyFlow {
+public class BorrowMoneyFlow {
 
   private long serialNumber;
   private long billId;
   private String inBoundAccount;
   private String outBoundAccount;
-  private java.sql.Timestamp exactDate;
   private double money;
-  private double interestMoney;
-  private double liquidatedMoney;
-  private double principalMoney;
+  private java.sql.Timestamp exactDate;
 
 
   public long getSerialNumber() {
@@ -50,15 +47,6 @@ public class RepayMoneyFlow {
   }
 
 
-  public java.sql.Timestamp getExactDate() {
-    return exactDate;
-  }
-
-  public void setExactDate(java.sql.Timestamp exactDate) {
-    this.exactDate = exactDate;
-  }
-
-
   public double getMoney() {
     return money;
   }
@@ -68,30 +56,12 @@ public class RepayMoneyFlow {
   }
 
 
-  public double getInterestMoney() {
-    return interestMoney;
+  public java.sql.Timestamp getExactDate() {
+    return exactDate;
   }
 
-  public void setInterestMoney(double interestMoney) {
-    this.interestMoney = interestMoney;
-  }
-
-
-  public double getLiquidatedMoney() {
-    return liquidatedMoney;
-  }
-
-  public void setLiquidatedMoney(double liquidatedMoney) {
-    this.liquidatedMoney = liquidatedMoney;
-  }
-
-
-  public double getPrincipalMoney() {
-    return principalMoney;
-  }
-
-  public void setPrincipalMoney(double principalMoney) {
-    this.principalMoney = principalMoney;
+  public void setExactDate(java.sql.Timestamp exactDate) {
+    this.exactDate = exactDate;
   }
 
 }
