+++ loan/src/main/java/loan/easyLoan/entity/LenderAccount.java
@@ -1,4 +1,4 @@
-package loan.easyloan.entity;
+package loan.easyLoan.entity;
 
 
 public class LenderAccount {
