+++ loan/src/main/resources/application.properties
@@ -1,0 +1,10 @@
+# \u6307\u5B9A\u65E5\u5FD7\u4F18\u5148\u7EA7
+logging.level.loan.easyLoan = trace
+
+# \u6307\u5B9A\u6587\u4EF6\u8DEF\u5F84/\u540D
+logging.file = log_test.log
+
+# \u6307\u5B9A\u63A7\u5236\u53F0\u8F93\u51FA\u65E5\u5FD7\u683C\u5F0F
+logging.pattern.console = %d{yyyy-mm-dd} [%thread] %-5level %logger{50} -%msg%n
+# \u6307\u5B9A\u6587\u4EF6\u4E2D\u7684\u65E5\u5FD7\u683C\u5F0F
+
