--- loan/src/test/java/p2ploan/loan/LoanApplicationTests.java
@@ -1,16 +1,0 @@
-package p2ploan.loan;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class LoanApplicationTests {
-
-    @Test
-    public void contextLoads() {
-    }
-
-}
